New class SendEmail in src/Service namespace
