+++ src/Controller/Controller.php
@@ -8,6 +8,7 @@ use BrunoGrosdidier\Blog\src\DAO\PostDAO;
 use BrunoGrosdidier\Blog\src\DAO\CommentDAO;
 use BrunoGrosdidier\Blog\src\DAO\UserDAO;
 use BrunoGrosdidier\Blog\src\Model\Pagination;
+use BrunoGrosdidier\Blog\src\Service\SendEmail;
 use BrunoGrosdidier\Blog\src\Service\Recaptcha;
 use Twig\Environment;
 use Twig\Extension\DebugExtension;
@@ -24,6 +25,7 @@ abstract class Controller
     protected $sentBySession;
     protected $validation;
     protected $pagination;
+    protected $sendEmail;
     protected $recaptcha;
     /**
      * @var Environment
@@ -36,6 +38,7 @@ abstract class Controller
         $this->commentDAO = new CommentDAO();
         $this->userDAO = new UserDAO();
         $this->pagination = new Pagination();
+        $this->sendEmail = new SendEmail();
         $this->recaptcha = new Recaptcha();
         $this->validation = new Validation();
         $request = new Request();
@@ -61,55 +64,5 @@ abstract class Controller
     public function render($template, $options = [])
     {
         echo $this->twig->render($template, $options);
-    }
-
-    public function sendEmailToMe($subjectToMe, $messageToMe)
-    {
-        // Create the Transport (Exp)
-        $transport = (new \Swift_SmtpTransport(EMAIL_HOST, EMAIL_PORT))
-            ->setUsername(EMAIL_USERNAME)
-            ->setPassword(EMAIL_PASSWORD)
-            ->setEncryption(EMAIL_ENCRYPTION) //For Gmail
-        ;
-
-        // Create the Mailer using your created Transport
-        $mailer = (new \Swift_Mailer($transport));
-
-        // Create a message (Dest)
-        $message = (new \Swift_Message($subjectToMe))
-            ->setFrom([EMAIL_USERNAME => EMAIL_SUBJECT])
-            ->setTo([EMAIL_DEST_1, EMAIL_DEST_2 => EMAIL_DEST_NAME])
-            ->setBody($messageToMe, 'text/html')
-        ;
-
-        // Send the message
-        $result = $mailer->send($message);
-
-        return $result;
-    }
-
-    public function sendEmailToUser($userEmail, $userName, $subjectToUser, $messageToUser)
-    {
-        // Create the Transport (Exp)
-        $transport = (new \Swift_SmtpTransport(EMAIL_HOST, EMAIL_PORT))
-            ->setUsername(EMAIL_USERNAME)
-            ->setPassword(EMAIL_PASSWORD)
-            ->setEncryption(EMAIL_ENCRYPTION) //For Gmail
-        ;
-
-        // Create the Mailer using your created Transport
-        $mailer = (new \Swift_Mailer($transport));
-
-        // Create a message (Dest)
-        $message = (new \Swift_Message($subjectToUser))
-            ->setFrom([EMAIL_USERNAME => EMAIL_SUBJECT])
-            ->setTo([$userEmail, $userEmail => $userName])
-            ->setBody($messageToUser, 'text/html')
-        ;
-
-        // Send the message
-        $result = $mailer->send($message);
-
-        return $result;
     }
 }
